Do not clean up spawners if told not to clean them
The documentation for passenger_max_preloader_idle_time indicates that
the spawner will never be cleaned up if it's set to '0'.  Setting it to
zero, though, causes Passenger to use the default setting of 300
seconds, since there was no check in the "gc idle spawner" code that the
max idle time is greater than zero.